Initial implementation of MkII
Changed network to only operate on a single dimension at a time; halves input nodes and forces X and Y logic to be the same (Good as the two dimensions share logic but are orthogonal. If they are both present the network is trying to learn to keep them the same as well as learning the problem)
Changed the minimum time before finding a target counts towards fitness from 1.5 to 5 seconds to stop rewarding luck when ball is falling off the edge regardless early on.
Changed the number of hidden nodes from 20 to 15 (less inputs so less hidden required). Genome length dropped from 200(?) to 125 -> smaller solution domain.